1、增加栏目链接TagHelpers 2、增加内容列表页和内容页控制器

+++ cms/CMS.Foundation/Modal/RequestModal/IPageRequest.cs
@@ -4,9 +4,9 @@ namespace Foundation.Modal.RequestModal
 {
     public interface IPageRequest
     {
-        long Current { get; set; }
+        int Current { get; set; }
 
-        long Size { get; set; }
+        int Size { get; set; }
 
         List<IQuery> Queries { get; set; }
